requirements.txt

pyGithub
